predicter/requirements.txt
pandas
numpy
https://download.pytorch.org/whl/cu100/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
torchvision
scikit-image
matplotlib